Set SESSION_COOKIE_AGE to an int

+++ dashboard_service/settings/common.py
@@ -184,4 +184,4 @@ if os.environ.get("SENTRY_DSN"):
 # Control Panel API settings
 CONTROL_PANEL_API_URL = os.environ.get("CONTROL_PANEL_API_URL")
 
-SESSION_COOKIE_AGE = os.environ.get("SESSION_COOKIE_AGE", 60 * 60 * 24 * 1)  # 1 day
+SESSION_COOKIE_AGE = int(os.environ.get("SESSION_COOKIE_AGE", 60 * 60 * 24 * 1))  # 1 day
